+++ VoxelizerX/Content/Voxelizer.h
@@ -25,8 +25,8 @@ public:
 	bool Init(XUSG::CommandList* pCommandList, uint32_t width, uint32_t height, XUSG::Format rtFormat,
 		XUSG::Format dsFormat, std::vector<XUSG::Resource>& uploaders, const char* fileName,
 		const DirectX::XMFLOAT4& posScale);
-	void UpdateFrame(uint32_t frameIndex, DirectX::CXMVECTOR eyePt, DirectX::CXMMATRIX viewProj);
-	void Render(const XUSG::CommandList* pCommandList, bool solid, Method voxMethod, uint32_t frameIndex,
+	void UpdateFrame(uint8_t frameIndex, DirectX::CXMVECTOR eyePt, DirectX::CXMMATRIX viewProj);
+	void Render(const XUSG::CommandList* pCommandList, bool solid, Method voxMethod, uint8_t frameIndex,
 		const XUSG::Descriptor& rtv, const XUSG::Descriptor& dsv);
 
 	static const uint8_t FrameCount = FRAME_COUNT;
@@ -137,13 +137,13 @@ protected:
 	bool prevoxelize(uint8_t mipLevel = 0);
 	bool prerenderBoxArray(XUSG::Format rtFormat, XUSG::Format dsFormat);
 	bool prerayCast(XUSG::Format rtFormat, XUSG::Format dsFormat);
-	void voxelize(const XUSG::CommandList* pCommandList, Method voxMethod, uint32_t frameIndex,
+	void voxelize(const XUSG::CommandList* pCommandList, Method voxMethod, uint8_t frameIndex,
 		bool depthPeel = false, uint8_t mipLevel = 0);
 	void voxelizeSolid(const XUSG::CommandList* pCommandList, Method voxMethod,
-		uint32_t frameIndex, uint8_t mipLevel = 0);
-	void renderBoxArray(const XUSG::CommandList* pCommandList, uint32_t frameIndex,
+		uint8_t frameIndex, uint8_t mipLevel = 0);
+	void renderBoxArray(const XUSG::CommandList* pCommandList, uint8_t frameIndex,
 		const XUSG::Descriptor& rtv, const XUSG::Descriptor& dsv);
-	void renderRayCast(const XUSG::CommandList* pCommandList, uint32_t frameIndex,
+	void renderRayCast(const XUSG::CommandList* pCommandList, uint8_t frameIndex,
 		const XUSG::Descriptor& rtv, const XUSG::Descriptor& dsv);
 
 	XUSG::Device m_device;
